Added functionality to vote to delete an incorrect scryfall response
Added a listener of discord.Reaction in addition to RawReactionActionEvent so it can count the number of reactions, and 4 or more ❌ reactions on a scryfall response deletes it. Useful in active chats to delete incorrect scryfall responses and unflood the chat, making backreading easier. Also its fun to collaboratively kill a message.